+++ packages/vue/CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2022-11-22
+
+### Fixed
+
+- InertiaFilterableList component now working properly with InertiaJS
+
+## [0.1.1] - 2022-11-22
+
+### Added
+
+- InertiaFilterableList component for usage with VueJS and InertiaJS
+
+## [0.1.0] - 2022-11-22
+
 ### Added
 
 - Initial release!